FEAT: Refactor HeaderComp, LinksNav, and SectionDots components to utilize layout store for state management and navigation
